Move type definitions to manager.go
